# Speed up float comparisons by avoiding branch mispredictions and boxing