--- a/bytecomp/translmod.ml
+++ b/bytecomp/translmod.ml
@@ -48,7 +48,7 @@ let field_path path field =
 let prim_set_oo_id =
   Pccall {Primitive.prim_name = "caml_set_oo_id"; prim_arity = 1;
           prim_alloc = false; prim_native_name = "";
-          prim_native_float = false}
+          prim_native_float = `No}
 
 let transl_extension_constructor env path ext =
   let name =