The current float compare suffers from [branch mispredictions](http://en.wikipedia.org/wiki/Branch_misprediction) and expects boxed floats.  This is noticeable for example in sorting float arrays.  This patch eliminates both problems.  This [benchmark](https://github.com/vbrankov/ocaml-demos/blob/master/fast-float-compare/test.ml) shows the following speedup on Intel Xeon E5-22687W:

 | No patch | With patch
---|------|------
Comparison easy to predict | 3.3 ns | 3 ns
Comparison random | 6.7 ns | 3 ns

OCaml accepts primitive C functions which accept and return unboxed floats.  Most of this patch addresses extending this feature to allow C functions to accept unboxed floats and return a value.  The rest of the patch is simply applying this extension on float comparison.

This [blog](https://blogs.janestreet.com/faster-ocaml-to-c-calls) contains another explanation of this problem.