@diml 2015-04-15
@vbrankov, I tried to generalize the idea of `"float"` and `"float-input"` to allow unboxing some of the arguments or the result of an external. The idea is to add `[@unbox_float]` on an argument type or on the result type to get it unboxed. It also supports `[@unbox_int32]`, `[@unbox_int64]` and `[@unbox_nativeint]`, which could help with [mantis 6551](http://caml.inria.fr/mantis/view.php?id=6551), for instance:

```ocaml
external bits_of_float : (float [@unbox_float]) -> (int64 [@unbox_int64]) = "caml_int64_bits_of_float"
```

I put the experiment here: https://github.com/diml/ocaml/tree/unbox_attributes_for_external

I tried it on 64 and 32 bits. I'll try to cleanup the patch and submit it.