@alainfrisch 2015-04-15
@diml : this is great, and I was about to suggest something along this direction.  Attributes seem to be the way forward to specify custom calling conventions on external declarations, and it makes sense to add granularity (each argument + result) and scope (int32/64) to the current unboxed calling convention.