@diml 2015-04-16
I submitted the patch in #167