@xavierleroy 2015-10-25
As far as I can see, there were 3 parts to this pull request:
(1) generalizing the "float" annotation over primitives to support additional forms of unboxing
(2) a clever branchless implementation of Pervasives.compare for float arguments
(3) using (1) to unbox the arguments to (2).

(1) was implemented (in a more general way) in #167 and merged in trunk.

I just cherry-picked (2) and put it in trunk (SVN commit 16539).  The branchless implementation generates pretty good code for x86-64 and for PowerPC, and decent (but far from optimal!) for ARM.

As for (3), I don't feel a strong need to unbox numerical arguments to Pervasives.compare.  In general, this primitive is used in the context of collections (Maps, Sets, etc), where the things being compared are generally boxed already.

With this, I'm closing this PR.