@alainfrisch 2015-10-26
@xavierleroy Would you be opposed to bringing back the unboxing part?  There could be uses in functions such as (float) array sorting or dichotomic lookup and this could become even more important when inlining is improved with flambda.